Barro Colorado Island (BCI) dataset

The BCI dataset contains location and species identity for all 10cm dbh (diameter at breast height) trees on Barro Colorado Island, currently for years 1981-1983, 1985, 1990, 1995, 2000, and 2005. The subset of interest here is the abundances for each of the 252 species recorded.

The BCI dataset is not included in the untb package, because its licence appears to be inconsistent with the GPL.

It is discussed here because it was used as an example dataset in Hankin 2007.

Source

http://www.stri.org/english/research/facilities/terrestrial/barro_colorado/index.php

References

R. Condit, S. P. Hubbell and R. B. Foster 2005. Barro Colorado Forest Census Plot Data

S. P. Hubbell 2001. “The Unified Neutral Theory of Biodiversity”. Princeton University Press.

R. K. S. Hankin 2007. “Introducing untb, an R package for simulating ecological drift under the unified neutral theory of biodiversity”. Journal of Statistical Software, volume 22, issue 12